refactor(favorite-locations): Move location components to subfolder.
