+++ lib/components/user/favorite-locations/favorite-location-controls.js
@@ -8,8 +8,8 @@ import React from 'react'
 import { DropdownButton, InputGroup, MenuItem } from 'react-bootstrap'
 import styled, { css } from 'styled-components'
 
-import connectLocationField from '../form/connect-location-field'
-import Icon from '../narrative/icon'
+import connectLocationField from '../../form/connect-location-field'
+import Icon from '../../narrative/icon'
 
 /**
  * This module contains components common to the favorite location components.
