+++ lib/components/user/favorite-locations/favorite-location.js
