+++ lib/components/user/favorite-locations/favorite-locations-pane.js
@@ -2,7 +2,7 @@ import { FieldArray } from 'formik'
 import React from 'react'
 import { ControlLabel } from 'react-bootstrap'
 
-import { isHomeOrWork } from '../../util/user'
+import { isHomeOrWork } from '../../../util/user'
 import FavoriteLocation from './favorite-location'
 import FixedFavoriteLocation from './fixed-favorite-location'
 import NewFavoriteLocation from './new-favorite-location'
