+++ lib/components/user/favorite-locations/fixed-favorite-location.js
@@ -6,7 +6,7 @@ import {
 } from 'react-bootstrap'
 import styled from 'styled-components'
 
-import Icon from '../narrative/icon'
+import Icon from '../../narrative/icon'
 import {
   FavoriteLocationField,
   FixedLocationType,
