+++ lib/components/user/favorite-locations/new-favorite-location.js
@@ -4,7 +4,7 @@ import {
   InputGroup
 } from 'react-bootstrap'
 
-import Icon from '../narrative/icon'
+import Icon from '../../narrative/icon'
 import {
   FavoriteLocationField,
   FIELD_HEIGHT_PX,
