+++ lib/components/user/user-account-screen.js
@@ -12,7 +12,7 @@ import { isNewUser, isHome, isWork } from '../../util/user'
 import DesktopNav from '../app/desktop-nav'
 import AccountSetupFinishPane from './account-setup-finish-pane'
 import ExistingAccountDisplay from './existing-account-display'
-import FavoriteLocationsPane from './favorite-locations-pane'
+import FavoriteLocationsPane from './favorite-locations/favorite-locations-pane'
 import NewAccountWizard from './new-account-wizard'
 import NotificationPrefsPane from './notification-prefs-pane'
 import TermsOfUsePane from './terms-of-use-pane'
